Fix circular dep; add usage to rendering
