+++ src/kaggle_benchmarks/__init__.py
@@ -29,9 +29,9 @@ from kaggle_benchmarks import (
 )
 from kaggle_benchmarks._config import ExecutionMode, config
 from kaggle_benchmarks.actors import Actor, LLMChat, system, user
-from kaggle_benchmarks.llm_messages import Usage
 from kaggle_benchmarks.runs import Run, Runs
 from kaggle_benchmarks.tasks import benchmark, task
+from kaggle_benchmarks.usage import Usage
 
 if kaggle.is_configured():
     llm = kaggle.load_default_model()
